requirements.txt
fastapi
sqlalchemy
sqlmodel
pydantic
bcrypt
jwt
stripe
python-dotenv
python-jose